+++ src/main/java/org/reactome/addlinks/fileprocessors/ZFINFileProcessor.java
@@ -1,5 +1,7 @@
 package org.reactome.addlinks.fileprocessors;
 
+import org.reactome.addlinks.EnsemblBiomartUtil;
+
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
@@ -29,7 +31,7 @@ public class ZFINFileProcessor extends FileProcessor{
         Path inputFilePath = Paths.get(this.pathToFile.toAbsolutePath().toString().replace(".gz", ""));
         List<String> lines = new ArrayList<>();
         try {
-            lines =  getLinesFromFile(inputFilePath, true);
+            lines = EnsemblBiomartUtil.getLinesFromFile(inputFilePath, true);
         } catch (IOException e) {
             logger.error("Error reading file ({}): {}", inputFilePath.toString(), e.getMessage());
             e.printStackTrace();
